Add several determinant algorithms

+++ Matrix/include/task.hpp
@@ -4,6 +4,6 @@
 
 namespace HWMatrix
 {
-Matrix<double> read_matrix(std::istream& input);
+Matrix<long double> read_matrix(std::istream& input);
 void task(std::istream& input, std::ostream& output);
 }
